+++ native_library/source/swift/PopupBehaviour.swift
@@ -20,15 +20,8 @@
 //  All Rights Reserved. Tua Rua Ltd.
 
 import Foundation
-
-public struct Constants {
-    public static let ON_URL_BLOCKED: String = "WebView.OnUrlBlocked"
-    public static let ON_POPUP_BLOCKED: String = "WebView.OnPopupBlocked"
-    public static let ON_FAIL: String = "WebView.OnFail"
-    public static let ON_ESC_KEY: String = "WebView.OnEscKey"
-    public static let ON_KEY_UP: String = "WebView.OnKeyUp"
-    public static let ON_KEY_DOWN: String = "WebView.OnKeyDown"
-    public static let ON_PROPERTY_CHANGE: String = "WebView.OnPropertyChange"
-    public static let JS_CALLBACK_EVENT: String = "TRWV.js.CALLBACK"
-    public static let AS_CALLBACK_EVENT: String = "TRWV.as.CALLBACK"
+public enum PopupBehaviour: Int {
+    case block = 0
+    case newWindow
+    case sameWindow
 }