+++ native_library/source/swift/SwiftController+WKUIDelegate.swift
@@ -20,15 +20,24 @@
 //  All Rights Reserved. Tua Rua Ltd.
 
 import Foundation
+import WebKit
 
-public struct Constants {
-    public static let ON_URL_BLOCKED: String = "WebView.OnUrlBlocked"
-    public static let ON_POPUP_BLOCKED: String = "WebView.OnPopupBlocked"
-    public static let ON_FAIL: String = "WebView.OnFail"
-    public static let ON_ESC_KEY: String = "WebView.OnEscKey"
-    public static let ON_KEY_UP: String = "WebView.OnKeyUp"
-    public static let ON_KEY_DOWN: String = "WebView.OnKeyDown"
-    public static let ON_PROPERTY_CHANGE: String = "WebView.OnPropertyChange"
-    public static let JS_CALLBACK_EVENT: String = "TRWV.js.CALLBACK"
-    public static let AS_CALLBACK_EVENT: String = "TRWV.as.CALLBACK"
+extension SwiftController: WKUIDelegate {
+  #if os(OSX)
+    // this handles <input type="file"/>
+    @available(OSX 10.12, *)
+    public func webView(_ webView: WKWebView, runOpenPanelWith parameters: WKOpenPanelParameters,
+                        initiatedByFrame frame: WKFrameInfo, completionHandler: @escaping ([URL]?) -> Void) {
+        let openPanel = NSOpenPanel()
+        openPanel.canChooseFiles = true
+        openPanel.allowsMultipleSelection = true
+        openPanel.begin(completionHandler: {(result) in
+            if result.rawValue == NSFileHandlingPanelOKButton {
+                completionHandler(openPanel.urls)
+            } else {
+                completionHandler([])
+            }
+        })
+    }
+    #endif
 }