+++ native_library/source/swift/SwiftController.swift
@@ -1,4 +1,4 @@
-// Copyright 2017 Tua Rua Ltd.
+// Copyright 2018 Tua Rua Ltd.
 // 
 //  Licensed under the Apache License, Version 2.0 (the "License");
 //  you may not use this file except in compliance with the License.
@@ -26,7 +26,7 @@ import FreSwift
 import Cocoa
 #endif
 
-public class SwiftController: NSObject, FreSwiftMainController, WKUIDelegate, WKNavigationDelegate, WKScriptMessageHandler {
+public class SwiftController: NSObject {
     public var TAG: String? = "WebViewANE"
     public var context: FreContextSwift!
     public var functionsToSet: FREFunctionMap = [:]
@@ -39,204 +39,52 @@ public class SwiftController: NSObject, FreSwiftMainController, WKUIDelegate, WK
     private static let zoomIncrement: CGFloat = CGFloat.init(0.1)
     private var _initialUrl: String = ""
     private var _viewPort: CGRect = CGRect.init(x: 0.0, y: 0.0, width: 800.0, height: 600.0)
-
-    public enum PopupBehaviour: Int {
-        case block = 0
-        case newWindow
-        case sameWindow
-    }
-
-    private var _popupBehaviour: PopupBehaviour = PopupBehaviour.newWindow;
+    internal var _popupBehaviour: PopupBehaviour = PopupBehaviour.newWindow
 #if os(iOS)
     private var _bgColor = UIColor.white
 #else
-    private var _popup: Popup!
+    internal var _popup: Popup!
 #endif
     private var _isAdded: Bool = false
-    private var _settings: Settings!
+    internal var _settings: Settings!
     private var _userController: WKUserContentController = WKUserContentController()
-
-    // must have this function !!
-    // Make sure these funcs match those in WebViewANE.m
-    @objc public func getFunctions(prefix: String) -> Array<String> {
-        functionsToSet["\(prefix)reload"] = reload
-        functionsToSet["\(prefix)load"] = load
-        functionsToSet["\(prefix)init"] = initWebView
-        functionsToSet["\(prefix)clearCache"] = clearCache
-        functionsToSet["\(prefix)isSupported"] = isSupported
-        functionsToSet["\(prefix)setVisible"] = setVisible
-        functionsToSet["\(prefix)loadHTMLString"] = loadHTMLString
-        functionsToSet["\(prefix)loadFileURL"] = loadFileURL
-        functionsToSet["\(prefix)onFullScreen"] = onFullScreen
-        functionsToSet["\(prefix)reloadFromOrigin"] = reloadFromOrigin
-        functionsToSet["\(prefix)stopLoading"] = stopLoading
-        functionsToSet["\(prefix)backForwardList"] = backForwardList
-        functionsToSet["\(prefix)go"] = go
-        functionsToSet["\(prefix)goBack"] = goBack
-        functionsToSet["\(prefix)goForward"] = goForward
-        functionsToSet["\(prefix)allowsMagnification"] = allowsMagnification
-        functionsToSet["\(prefix)zoomIn"] = zoomIn
-        functionsToSet["\(prefix)zoomOut"] = zoomOut
-        functionsToSet["\(prefix)setViewPort"] = setViewPort
-        functionsToSet["\(prefix)showDevTools"] = showDevTools
-        functionsToSet["\(prefix)closeDevTools"] = closeDevTools
-        functionsToSet["\(prefix)callJavascriptFunction"] = callJavascriptFunction
-        functionsToSet["\(prefix)evaluateJavaScript"] = evaluateJavaScript
-        functionsToSet["\(prefix)injectScript"] = injectScript
-        functionsToSet["\(prefix)focus"] = focusWebView
-        functionsToSet["\(prefix)print"] = print
-        functionsToSet["\(prefix)capture"] = capture
-        functionsToSet["\(prefix)addTab"] = addTab
-        functionsToSet["\(prefix)closeTab"] = closeTab
-        functionsToSet["\(prefix)setCurrentTab"] = setCurrentTab
-        functionsToSet["\(prefix)getCurrentTab"] = getCurrentTab
-        functionsToSet["\(prefix)getTabDetails"] = getTabDetails
-        functionsToSet["\(prefix)shutDown"] = shutDown
-        functionsToSet["\(prefix)addEventListener"] = addEventListener
-        functionsToSet["\(prefix)removeEventListener"] = removeEventListener
-
-        var arr: Array<String> = []
-        for key in functionsToSet.keys {
-            arr.append(key)
-        }
-
-        return arr
-    }
-#if os(OSX)
-    // this handles <input type="file"/>
-    @available(OSX 10.12, *)
-    public func webView(_ webView: WKWebView, runOpenPanelWith parameters: WKOpenPanelParameters, initiatedByFrame frame: WKFrameInfo, completionHandler: @escaping ([URL]?) -> Void) {
-        let openPanel = NSOpenPanel()
-        openPanel.canChooseFiles = true
-        openPanel.allowsMultipleSelection = true
-        openPanel.begin(completionHandler: {(result) in
-            if result.rawValue == NSFileHandlingPanelOKButton {
-                completionHandler(openPanel.urls)
-            }else{
-                completionHandler([])
-            }
-        })
-    }
-#endif
-    public func webView(_ webView: WKWebView, didReceive challenge: URLAuthenticationChallenge, completionHandler: @escaping (URLSession.AuthChallengeDisposition, URLCredential?) -> Void) {
-        guard let serverTrust = challenge.protectionSpace.serverTrust else { return completionHandler(.useCredential, nil) }
-        let exceptions = SecTrustCopyExceptions(serverTrust)
-        SecTrustSetExceptions(serverTrust, exceptions)
-        completionHandler(.useCredential, URLCredential(trust: serverTrust))
-    }
     
-    // this handles target=_blank links by opening them in the same view
-    public func webView(_ webView: WKWebView, createWebViewWith configuration: WKWebViewConfiguration,
-                 for navigationAction: WKNavigationAction, windowFeatures: WKWindowFeatures) -> WKWebView? {
-        if navigationAction.targetFrame == nil {
-            switch _popupBehaviour {
-            case .block:
-                var props: Dictionary<String, Any> = Dictionary()
-                props["url"] = ""
-                if let url = navigationAction.request.url?.absoluteString {
-                    props["url"] = url
-                }
-                props["tab"] = getCurrentTab(webView)
-                let json = JSON(props)
-                sendEvent(name: Constants.ON_POPUP_BLOCKED, value: json.description)
-                break
-            case .newWindow:
-#if os(iOS)
-                warning("Cannot open popup in new window on iOS. Opening in same window.")
-                webView.load(navigationAction.request)
-#else
-                _popup.createPopupWindow(url: navigationAction.request, configuration: _settings.configuration)
-#endif
-                break
-            case .sameWindow:
-                webView.load(navigationAction.request)
-                break
-            }
-        }
-        return nil
-    }
-
-
-    public func webView(_ webView: WKWebView, decidePolicyFor navigationAction: WKNavigationAction, decisionHandler: @escaping (WKNavigationActionPolicy) -> Void) {
-        guard let newUrl = navigationAction.request.url?.absoluteString.lowercased()
-          else {
-            decisionHandler(.allow)
-            return
-        }
-
-        if isWhiteListBlocked(url: newUrl) || isBlackListBlocked(url: newUrl) {
-            var props: Dictionary<String, Any> = Dictionary()
-            props["url"] = newUrl
-            props["tab"] = getCurrentTab(webView)
-
-            let json = JSON(props)
-            sendEvent(name: Constants.ON_URL_BLOCKED, value: json.description)
-            decisionHandler(.cancel)
-        } else {
-            decisionHandler(.allow)
-        }
-
-    }
-
-    public func webView(_ webView: WKWebView, didFail navigation: WKNavigation!, withError: Error) {
-        var props: Dictionary<String, Any> = Dictionary()
-        props["url"] = webView.url!.absoluteString
-        props["errorCode"] = 0
-        props["errorText"] = withError.localizedDescription
-        let json = JSON(props)
-        sendEvent(name: Constants.ON_FAIL, value: json.description)
-    }
-
     private func addKeyListener(type: String) {
 #if os(OSX)
         let eventMask: NSEvent.EventTypeMask = type == "keyUp" ? .keyUp : .keyDown
         var listener: Any? = type == "keyUp" ? SwiftController.keyUpListener : SwiftController.keyDownListener
-        if (listener == nil) {
+        if listener == nil {
             listener = NSEvent.addLocalMonitorForEvents(matching: [eventMask]) { (event: NSEvent) -> NSEvent? in
                 var modifiers: String = ""
                 switch event.modifierFlags.intersection(NSEvent.ModifierFlags.deviceIndependentFlagsMask) {
                 case [.shift]:
                     modifiers = "shift"
-                    break
                 case [.control]:
                     modifiers = "control"
-                    break
                 case [.option]:
                     modifiers = "alt"
-                    break
                 case [.command]:
                     modifiers = "command"
-                    break
                 case [.control, .shift]:
                     modifiers = "control-shift"
-                    break
                 case [.option, .shift]:
                     modifiers = "option-shift"
-                    break
                 case [.command, .shift]:
                     modifiers = "command-shift"
-                    break
                 case [.control, .shift]:
                     modifiers = "control-shift"
-                    break
                 case [.control, .command]:
                     modifiers = "control-command"
-                    break
                 case [.option, .command]:
                     modifiers = "alt-command"
-                    break
                 case [.shift, .control, .option]:
                     modifiers = "shift-control-option"
-                    break
                 case [.shift, .control, .command]:
                     modifiers = "shift-control-command"
-                    break
                 case [.shift, .option, .command]:
                     modifiers = "shift-alt-command"
-                    break
                 case [.shift, .command, .option]:
                     modifiers = "shift-command-alt"
-                    break
                 default:
                     break
                 }
@@ -247,33 +95,27 @@ public class SwiftController: NSObject, FreSwiftMainController, WKUIDelegate, WK
                     switch event.keyCode {
                     case 51: //BACKSPACE
                         keyValue = 8
-                        break
                     case 53: //ESC
                         keyValue = 27
-                        break
                     case 123: //LEFT
                         keyValue = 37
-                        break
                     case 124: //RIGHT
                         keyValue = 39
-                        break
                     case 125: //DOWN
                         keyValue = 40
-                        break
                     case 126: //UP
                         keyValue = 38
-                        break
                     default:
                         keyValue = s[s.startIndex].value
-                        break
                     }
-                    var props: Dictionary<String, Any> = Dictionary()
+                    var props: [String: Any] = Dictionary()
                     props["keyCode"] = keyValue
                     props["nativeKeyCode"] = event.keyCode
                     props["modifiers"] = modifiers
                     props["isSystemKey"] = false
                     let json = JSON(props)
-                    self.sendEvent(name: event.type == .keyUp ? Constants.ON_KEY_UP : Constants.ON_KEY_DOWN, value: json.description)
+                    self.sendEvent(name: event.type == .keyUp ? Constants.ON_KEY_UP : Constants.ON_KEY_DOWN,
+                                   value: json.description)
 
                 }
 
@@ -416,7 +258,7 @@ public class SwiftController: NSObject, FreSwiftMainController, WKUIDelegate, WK
             return ArgCountError(message: "setVisible").getError(#file, #line, #column)
         }
 
-        if (!_isAdded) {
+        if !_isAdded {
             addToStage()
         }
         wv.isHidden = !visible
@@ -515,29 +357,33 @@ public class SwiftController: NSObject, FreSwiftMainController, WKUIDelegate, WK
 
         let tmpIsAdded = _isAdded
         for win in NSApp.windows {
-            if (fullScreen && win.canBecomeMain && win.className.contains("AIR_FullScreen")) {
+            if fullScreen && win.canBecomeMain && win.className.contains("AIR_FullScreen") {
                 win.makeMain()
-                if (SwiftController.escListener == nil) {
-                    SwiftController.escListener = NSEvent.addLocalMonitorForEvents(matching: [NSEvent.EventTypeMask.keyUp]) { (event: NSEvent) -> NSEvent? in
+                if SwiftController.escListener == nil {
+                    SwiftController.escListener = NSEvent.addLocalMonitorForEvents(
+                    matching: [NSEvent.EventTypeMask.keyUp]) { (event: NSEvent) -> NSEvent? in
                         let theX = event.locationInWindow.x
                         let theY = event.locationInWindow.y
-                        let realY = ((NSApp.mainWindow?.contentLayoutRect.height)! - self._viewPort.size.height) - self._viewPort.origin.y
-                        if (event.keyCode == 53 && theX > self._viewPort.origin.x && theX < (self._viewPort.size.width - self._viewPort.origin.x)
-                          && theY > realY && theY < (realY + self._viewPort.size.height)) {
+                        let realY = ((NSApp.mainWindow?.contentLayoutRect.height)!
+                            - self._viewPort.size.height) - self._viewPort.origin.y
+                        if event.keyCode == 53
+                            && theX > self._viewPort.origin.x
+                            && theX < (self._viewPort.size.width - self._viewPort.origin.x)
+                            && theY > realY && theY < (realY + self._viewPort.size.height) {
                             self.sendEvent(name: Constants.ON_ESC_KEY, value: "")
                         }
                         return event
                     }
                 }
                 break
-            } else if (!fullScreen && win.canBecomeMain && win.className.contains("AIR_PlayerContent")) {
+            } else if !fullScreen && win.canBecomeMain && win.className.contains("AIR_PlayerContent") {
                 win.makeMain()
                 win.orderFront(nil)
                 break
             }
         }
 
-        if (tmpIsAdded) {
+        if tmpIsAdded {
             _ = removeFromStage()
             _ = addToStage()
         }
@@ -644,8 +490,8 @@ public class SwiftController: NSObject, FreSwiftMainController, WKUIDelegate, WK
             let w = wFre * Int(UIScreen.main.scale)
             let h = hFre * Int(UIScreen.main.scale)
             do {
-                if let freObject = try FREObject.init(className: "flash.display.BitmapData", args: cg.width, cg.height, false)
-                     {
+                if let freObject = try FREObject.init(className: "flash.display.BitmapData",
+                                                      args: cg.width, cg.height, false) {
                     let asBitmapData = FreBitmapDataSwift.init(freObject: freObject)
                     defer {
                         asBitmapData.releaseData()
@@ -654,11 +500,14 @@ public class SwiftController: NSObject, FreSwiftMainController, WKUIDelegate, WK
                         try asBitmapData.acquire()
                         try asBitmapData.setPixels(cgImage: cg)
                         asBitmapData.releaseData()
-                        if (w > 0 && h > 0) {
-                            if let destBmd = try FREObject.init(className: "flash.display.BitmapData", args: w, h, false) {
+                        if w > 0 && h > 0 {
+                            if let destBmd = try FREObject.init(className: "flash.display.BitmapData",
+                                                                args: w, h, false) {
                                 let rect = FreRectangleSwift.init(value: CGRect.init(x: x, y: y, width: w, height: h))
                                 let pt = FrePointSwift.init(value: CGPoint.zero)
-                                if let bmd = asBitmapData.rawValue, let sourceRect = rect.rawValue, let destPoint = pt.rawValue {
+                                if let bmd = asBitmapData.rawValue,
+                                    let sourceRect = rect.rawValue,
+                                    let destPoint = pt.rawValue {
                                     _ = try destBmd.call(method: "copyPixels", args: bmd, sourceRect, destPoint)
                                     return destBmd
                                 }
@@ -696,8 +545,12 @@ public class SwiftController: NSObject, FreSwiftMainController, WKUIDelegate, WK
     func clearCache(ctx: FREContext, argc: FREArgc, argv: FREArgv) -> FREObject? {
         if #available(iOS 9.0, OSX 10.11, *) {
             let websiteDataTypes = NSSet(array: [WKWebsiteDataTypeDiskCache, WKWebsiteDataTypeMemoryCache])
-            WKWebsiteDataStore.default().removeData(ofTypes: websiteDataTypes as! Set<String>, modifiedSince: NSDate(timeIntervalSince1970: 0) as Date, completionHandler: {})
-
+            if let websiteDataTypes = websiteDataTypes as? Set<String> {
+                WKWebsiteDataStore.default().removeData(
+                    ofTypes: websiteDataTypes,
+                    modifiedSince: NSDate(timeIntervalSince1970: 0) as Date,
+                    completionHandler: {})
+            }
         }
         return nil
     }
@@ -716,10 +569,11 @@ public class SwiftController: NSObject, FreSwiftMainController, WKUIDelegate, WK
             _initialUrl = ""
         }
 
-        _currentTab = _tabList.count;
+        _currentTab = _tabList.count
 
         _currentWebView = createNewBrowser(
-          frame: CGRect.init(x: wv.frame.origin.x, y: wv.frame.origin.y, width: _viewPort.size.width, height: _viewPort.size.height),
+          frame: CGRect.init(x: wv.frame.origin.x, y: wv.frame.origin.y,
+                             width: _viewPort.size.width, height: _viewPort.size.height),
           tab: _tabList.count)
 
         _ = removeFromStage()
@@ -742,7 +596,7 @@ public class SwiftController: NSObject, FreSwiftMainController, WKUIDelegate, WK
 
         let doRefresh = (_currentTab >= index)
         if _currentTab >= index {
-            _currentTab = _currentTab - 1
+            _currentTab -= 1
         }
         if _tabList.count == 2 {
             _currentTab = 0
@@ -750,7 +604,6 @@ public class SwiftController: NSObject, FreSwiftMainController, WKUIDelegate, WK
         if _currentTab < 0 {
             _currentTab = 0
         }
-
 
         var wvtc: WebViewVC? = _tabList.object(at: index) as? WebViewVC
         if let wvToClose = wvtc {
@@ -761,7 +614,7 @@ public class SwiftController: NSObject, FreSwiftMainController, WKUIDelegate, WK
         wvtc = nil
 
         let currentFrame = cwv.frame
-        _currentWebView = _tabList[_currentTab] as? WebViewVC;
+        _currentWebView = _tabList[_currentTab] as? WebViewVC
 
         var cnt = 0
         for vc in _tabList {
@@ -769,10 +622,10 @@ public class SwiftController: NSObject, FreSwiftMainController, WKUIDelegate, WK
                 theVC.tab = cnt
                 theVC.isHidden = !(theVC.tab == _currentTab)
             }
-            cnt = cnt + 1
+            cnt += 1
         }
 
-        if (!doRefresh) {
+        if !doRefresh {
             return nil
         }
 
@@ -787,7 +640,6 @@ public class SwiftController: NSObject, FreSwiftMainController, WKUIDelegate, WK
 #endif
         return nil
     }
-
 
     func setCurrentTab(ctx: FREContext, argc: FREArgc, argv: FREArgv) -> FREObject? {
 #if os(OSX)
@@ -807,9 +659,8 @@ public class SwiftController: NSObject, FreSwiftMainController, WKUIDelegate, WK
             }
         }
         let currentFrame = cwv.frame
-        _currentTab = index;
-        _currentWebView = _tabList[_currentTab] as? WebViewVC;
-
+        _currentTab = index
+        _currentWebView = _tabList[_currentTab] as? WebViewVC
 
         guard let wv = _currentWebView else {
             return nil
@@ -830,10 +681,17 @@ public class SwiftController: NSObject, FreSwiftMainController, WKUIDelegate, WK
             var cnt = 0
             for vc in _tabList {
                 if let theVC = vc as? WebViewVC {
-                    if let currentTabFre = try FREObject.init(className: "com.tuarua.webview.TabDetails", args: theVC.tab, theVC.url!.absoluteString, theVC.title!, theVC.isLoading, theVC.canGoBack, theVC.canGoForward, theVC.estimatedProgress) {
+                    if let currentTabFre = try FREObject.init(className: "com.tuarua.webview.TabDetails",
+                                                              args: theVC.tab,
+                                                              theVC.url!.absoluteString,
+                                                              theVC.title!,
+                                                              theVC.isLoading,
+                                                              theVC.canGoBack,
+                                                              theVC.canGoForward,
+                                                              theVC.estimatedProgress) {
                         try airArray.set(index: UInt(cnt), value: currentTabFre)
                     }
-                    cnt = cnt + 1
+                    cnt += 1
                 }
             }
         } catch {
@@ -844,27 +702,6 @@ public class SwiftController: NSObject, FreSwiftMainController, WKUIDelegate, WK
     func getCurrentTab(ctx: FREContext, argc: FREArgc, argv: FREArgv) -> FREObject? {
         return _currentTab.toFREObject()
     }
-
-#if os(iOS)
-
-    public func userContentController(_ userContentController: WKUserContentController, didReceive message: WKScriptMessage) {
-        if let messageBody: NSDictionary = message.body as? NSDictionary {
-            let json = JSON(messageBody)
-            sendEvent(name: Constants.JS_CALLBACK_EVENT, value: json.description)
-        }
-    }
-
-#else
-
-    @available(OSX 10.10, *)
-    public func userContentController(_ userContentController: WKUserContentController, didReceive message: WKScriptMessage) {
-        if let messageBody: NSDictionary = message.body as? NSDictionary {
-            let json = JSON(messageBody)
-            sendEvent(name: Constants.JS_CALLBACK_EVENT, value: json.description)
-        }
-    }
-
-#endif
 
     fileprivate func createNewBrowser(frame: CGRect, tab: Int) -> WebViewVC {
         let wv = WebViewVC(context: context, frame: frame, configuration: _settings.configuration, tab: tab)
@@ -899,7 +736,7 @@ public class SwiftController: NSObject, FreSwiftMainController, WKUIDelegate, WK
         return wv
     }
 
-    func initWebView(ctx: FREContext, argc: FREArgc, argv: FREArgv) -> FREObject? {
+    func initController(ctx: FREContext, argc: FREArgc, argv: FREArgv) -> FREObject? {
         guard argc > 4,
               let inFRE4 = argv[4],
               let viewPortFre = CGRect(argv[1])
@@ -928,25 +765,25 @@ public class SwiftController: NSObject, FreSwiftMainController, WKUIDelegate, WK
                 mWin = allWindows[0]
             }
         } else {
-            return FreError(stackTrace: "", message: "Cannot find AIR window to attach webView to. Ensure you init the ANE AFTER your main Sprite is initialised. " +
-              "Please see https://forum.starling-framework.org/topic/webviewane-for-osx/page/7?replies=201#post-105524 for more details", type: FreError.Code.ok).getError(#file, #line, #column)
+            return FreError(stackTrace: "", message: "Cannot find AIR window to attach webView to. Ensure you init "
+                + "the ANE AFTER your main Sprite is initialised. Please see " +
+              "https://forum.starling-framework.org/topic/webviewane-for-osx/page/7?replies=201#post-105524 " +
+                "for more details", type: FreError.Code.ok).getError(#file, #line, #column)
         }
 
         realY = ((mWin?.contentLayoutRect.height)! - _viewPort.size.height) - _viewPort.origin.y
 #endif
 
-
         if let settingsFRE: FREObject = argv[2] {
-            if let settingsDict:Dictionary<String, AnyObject> = Dictionary.init(settingsFRE) {
+            if let settingsDict: [String: AnyObject] = Dictionary.init(settingsFRE) {
                 _settings = Settings.init(dictionary: settingsDict)
-
 #if os(OSX)
                 if let popupSettings = settingsDict["popup"] {
                     _popup = Popup()
                     if let behaviour = popupSettings["behaviour"] as? Int {
-                        _popupBehaviour = SwiftController.PopupBehaviour(rawValue: behaviour)!
+                        _popupBehaviour = PopupBehaviour(rawValue: behaviour)!
                     }
-                    if let dimensions = popupSettings["dimensions"] as? Dictionary<String, Int> {
+                    if let dimensions = popupSettings["dimensions"] as? [String: Int] {
                         if let w = dimensions["width"], let h = dimensions["height"] {
                             _popup.popupDimensions.0 = w
                             _popup.popupDimensions.1 = h
@@ -960,13 +797,14 @@ public class SwiftController: NSObject, FreSwiftMainController, WKUIDelegate, WK
         _userController.add(self, name: "webViewANE")
         _settings.configuration.userContentController = _userController
 
-        _currentWebView = createNewBrowser(frame: CGRect.init(x: _viewPort.origin.x, y: realY, width: _viewPort.size.width, height: _viewPort.size.height), tab: 0)
+        _currentWebView = createNewBrowser(frame:
+            CGRect(x: _viewPort.origin.x, y: realY, width: _viewPort.size.width, height: _viewPort.size.height),
+                                           tab: 0)
 
         return nil
     }
 
-
-    fileprivate func getCurrentTab(_ webView: WKWebView?) -> Int {
+    internal func getCurrentTab(_ webView: WKWebView?) -> Int {
         if let wv = webView as? WebViewVC {
             for vc in _tabList {
                 if let theVC = vc as? WebViewVC {
@@ -979,21 +817,21 @@ public class SwiftController: NSObject, FreSwiftMainController, WKUIDelegate, WK
         return 0
     }
 
-    fileprivate func isWhiteListBlocked(url: String) -> Bool {
+    internal func isWhiteListBlocked(url: String) -> Bool {
         guard let list = _settings.urlWhiteList,
               list.count > 0
           else {
             return false
         }
         for item in list {
-            if url.range(of: item as! String) != nil {
+            if let item = item as? String, url.range(of: item) != nil {
                 return false
             }
         }
         return true
     }
 
-    fileprivate func isBlackListBlocked(url: String) -> Bool {
+    internal func isBlackListBlocked(url: String) -> Bool {
         guard let list = _settings.urlBlackList,
               list.count > 0
           else {
@@ -1001,28 +839,12 @@ public class SwiftController: NSObject, FreSwiftMainController, WKUIDelegate, WK
         }
 
         for item in list {
-            if url.range(of: item as! String) != nil {
+            if let item = item as? String, url.range(of: item) != nil {
                 return true
             }
         }
 
         return false
-    }
-
-
-    // Must have this function. It exposes the methods to our entry ObjC.
-    @objc public func callSwiftFunction(name: String, ctx: FREContext, argc: FREArgc, argv: FREArgv) -> FREObject? {
-        if let fm = functionsToSet[name] {
-            return fm(ctx, argc, argv)
-        }
-        return nil
-    }
-    
-    @objc public func setFREContext(ctx: FREContext) {
-        self.context = FreContextSwift.init(freContext: ctx)
-    }
-    
-    @objc public func onLoad() {
     }
 
 }